Implement drag and drop list management
This commit brings the ability to drag and drop staff around multiple
lists. There is still currently no way to add staff other that your
browser's console, but once you add some you can drag and drop them
around lists. One thing to note is that I didn't put code in for
maintaining list position, as it adds complexity and data without very
much benefit. You may notice after dropping an item the item jumps to
elsewhere on the list, since that's where meteor thinks it should be. I
don't think it's that big of a deal.

This commit also changes the `null` value for lists to be `"master"`, as
nulls just added complexity for no real good reason. Might as well have
everything be a string. `.gitignore` also updated to not track Vim swap
files.